requirements.txt
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
firebase-admin==4.5.2
Flask==1.1.2
Flask-Cors==3.0.10
gcloud==0.17.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jws==0.1.3
MarkupSafe==1.1.1
msgpack==1.0.2
oauth2client==3.0.0
packaging==20.9
proto-plus==1.13.0
protobuf==3.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyparsing==2.4.7
python-dotenv==0.15.0
python-jwt==2.0.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.7.1
six==1.15.0
stripe==2.55.2
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
